add USE_SDL define for the upcoming SDL port
In USE_SDL mode, most things will work as in the Unix
ports, so that part is simple. However, the WIN32 define
will also be set when compiling on Windows, and we don't
want the Win32-specific code that's used in some part of
the demo to interfere with that. As a result, lots of
ifdefs need to be changed.

This commit also includes a dummy implementation of an
SDL-specific timer module, and adds two missing #includes.