avoid use of uninitialized variables
In the LZW decoder, there are two variables that are
declared in an inner loop, but shall be preserved
across iterations. Newer compilers (rightfully) complain
about them being uninitialized, so let's correct all that.

Fun fact: An earlier iteration of this fix did *not*
move the declaration and initialization to where it
belonged, but re-initialized the variables with each
iteration, causing all LZW decoding to be broken ...
and thus costing me two extra days of debugging :/