initial draft of an SDL2 + OpenGL graphics output driver
Still in rough shape, but it works ...
It's video-only at the moment, so sound is turned off
for the time being.

SDL2 is used for window management and timing,
an OpenGL texture is used for rendering.
(SDL_Renderer ist *not* used because it can only
do nearest-neighbor or bilinear interpolation,
but I have a fancy "antialiased blocky" shader
in planning, so I need to use OpenGL directly.)

The OpenGL loader is fully custom to avoid
issues with different linkage of extension
functions on different platforms.