crashing bug workaround
For unknown reasons, the 64-bit builds semi-reliably crash
in the "blob" / metaballs / marching cubes scene half-way
into the demo. It's nearly reproducible on all systems and
configurations (Debug/Release), but it's not always at the
same place, and it doesn't happen when starting the demo
at that exact scene or the one before it. Exactly the kind
of bug you don't want!

Whatever the root cause is, the symptom is that a division
by zero can occur deep in the bowels of the marching cubes
algorithm, causing a subsequent table look-up to fail
(because it's looking up an "infinite" address).
A simple band-aid that just clips the result of the division
to zero if the denominator is zero seems to do the trick.
I didn't observe any graphical glitches so far.