work around vista scene render bugs
The canyon scene had a few severe rendering issues with some objects,
especially the spaceships: some polygons were just missing.

I couldn't find the exact root cause, but I found the culprit:
the object simplification ("optimization") that tries to reduce
vertex counts. If that is disabled, along with the vista scene
loading optimization (which depends on the vertex IDs being exactly
the same as expected), the geometry is (mostly?) bug-free now.