automatic fallback to OpenGL 1.x
OpenGL 3.3 is a pretty solid bet for desktop PCs nowadays,
but there are still platforms with less capable drivers,
most notably the Raspberry Pi.

To make the demo run on such systems too, the display code
now uses whatever version is available, and if it's 3.3
or later, the shader-based pipeline is used. If it's older,
it falls back to OpenGL 1.1 and glBegin()/glVertex()/glEnd().

This means that we no longer run in Core Profile, because
switching between profiles is a lot of hassle. (The
window needs to be destroyed as well to switch profiles.)
Since there's no strict reason to use Core Profile other
than moral high ground, we accept falling back to
Compatibility Profile for now.