remove (almost) all #pragma packs
There are only very few (namely, the ones which are used to read
structures from the data file) which really need packing, and now
that we got a good understanding of the source code and where this
situation applies, we can remove the packing from most other
structures. This avoids issues on architectures with stricter
alignment rules (ARMv6/v7, mainly).